server/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-restful
python-dotenv
marshmallow
flask-admin
flask-cors
flask-login